+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project are documented in this file, based on [Keep 
 
 
 ## [Unreleased]
+### Fixed
+- Plugin can now be used without applying `maven-publish`.
 
 
 ## [1.0.0] - 2024-05-08